modules/inference_pipeline/requirements.txt
gradio==4.42.0
gradio_client==1.3.0
nltk==3.8.1
kafka-python==2.0.2
ollama==0.2.1
pydantic==2.8.0
pydantic_core==2.20.0